+++ keteparaha/__init__.py
@@ -13,4 +13,4 @@
     :license: MIT, see LICENSE for more details
 """
 
-__version__ = '0.0.8'
+__version__ = '0.0.9'